+++ cs/Markdown/Parser/EventType.cs
@@ -1,0 +1,13 @@
+﻿namespace Markdown.Parser
+{
+    public enum EventType
+    {
+        PlainText = 0,
+
+        BoldStart,
+        BoldEnd,
+
+        ItalicStart,
+        ItalicEnd
+    }
+}
